refactor file structure
